+++ extend/bajo/hook/dobo.sumba-route-guard@after-transaction.js
@@ -1,0 +1,6 @@
+async function afterTransaction (action, ...args) {
+  if (!['createRecord', 'updateRecord', 'removeRecord'].includes(action)) return
+  await this.getRouteGuards(true)
+}
+
+export default afterTransaction