+++ extend/bajo/hook/dobo.sumba-team-guard@after-action.js
@@ -1,6 +1,0 @@
-async function afterAction (action, ...args) {
-  if (!['createRecord', 'updateRecord', 'removeRecord'].includes(action)) return
-  await this.getTeamGuards(true)
-}
-
-export default afterAction