+++ extend/dobo/model/route-guard.js
@@ -1,9 +1,20 @@
 async function routeGuard () {
   return {
-    connection: 'memory',
     properties: [
       'path,,255,true,true',
       {
+        name: 'methods',
+        type: 'array',
+        required: true,
+        default: ['GET', 'POST', 'UPDATE', 'DELETE'],
+        values: ['GET', 'POST', 'UPDATE', 'DELETE']
+      },
+      'teamIds,sumba:teamIds',
+      {
+        name: 'weight',
+        type: 'smallint',
+        default: 0
+      }, {
         name: 'inverse',
         type: 'boolean',
         required: true,
@@ -13,28 +24,6 @@ async function routeGuard () {
         type: 'boolean',
         required: true,
         default: false
-      }, {
-        name: 'methods',
-        type: 'array',
-        required: true,
-        default: ['GET', 'POST', 'UPDATE', 'DELETE'],
-        values: ['GET', 'POST', 'UPDATE', 'DELETE']
-      }, {
-        name: 'teams',
-        type: 'array',
-        default: [],
-        ref: {
-          teams: {
-            model: 'SumbaTeam',
-            field: 'alias',
-            searchField: 'name',
-            fields: ['alias', 'name']
-          }
-        }
-      }, {
-        name: 'weight',
-        type: 'smallint',
-        default: 0
       }
     ],
     features: [
@@ -43,16 +32,12 @@ async function routeGuard () {
         values: ['ACTIVE', 'INACTIVE'],
         default: 'ACTIVE'
       },
-      {
-        name: 'dobo:unique',
-        fields: ['path', 'anonymous', 'methods', 'teams', 'inverse', 'weight', 'status', 'siteId']
-      },
       'dobo:immutable',
       'dobo:updatedAt',
       'sumba:siteId'
     ],
     options: {
-      persistence: false
+      attachment: false
     }
   }
 }