+++ extend/waibuMpa/extend/waibuAdmin/route/model-guard/@action.js
@@ -1,12 +1,10 @@
 const action = {
   method: ['GET', 'POST'],
-  title: 'cacheStorage',
-  crossSite: true,
+  title: 'modelGuard',
   handler: async function (req, reply) {
-    if (!this.app.bajoCache) throw this.error('_notFound')
     const { importModule } = this.app.bajo
     const crudSkel = await importModule('waibuAdmin:/lib/crud-skel.js')
-    return await crudSkel.call(this, 'CacheStorage', req, reply)
+    return await crudSkel.call(this, 'SumbaModelGuard', req, reply)
   }
 }