+++ extend/waibuMpa/extend/waibuAdmin/route/session/@action.js
@@ -1,7 +1,7 @@
 const action = {
   method: ['GET', 'POST'],
   title: 'userSession',
-  interSite: true,
+  crossSite: true,
   handler: async function (req, reply) {
     const { importModule } = this.app.bajo
     const crudSkel = await importModule('waibuAdmin:/lib/crud-skel.js')